fasd

+++ service/src/main/java/com/wangda/alarm/service/bean/standard/alarminfo/alarm/AlarmLevel.java
@@ -5,19 +5,19 @@ package com.wangda.alarm.service.bean.standard.alarminfo.alarm;
  * @version 2017-10-23
  */
 public enum  AlarmLevel {
-    LEVEL_ONE(1, "1级报警"),
-    LEVEL_TWO(2, "2级报警"),
-    LEVEL_THREE(3, "3级报警"),
-    WARN(4, "预警");
-    private int code;
+    LEVEL_ONE((byte) 1, "1级报警"),
+    LEVEL_TWO((byte)2, "2级报警"),
+    LEVEL_THREE((byte)3, "3级报警"),
+    WARN((byte)4, "预警");
+    private byte code;
     private String desc;
 
-    AlarmLevel(int code, String desc) {
+    AlarmLevel(byte code, String desc) {
         this.code = code;
         this.desc = desc;
     }
 
-    public static AlarmLevel codeOf(int code) {
+    public static AlarmLevel codeOf(byte code) {
         for (AlarmLevel level : values()) {
             if (level.code == code)
                 return level;
